+++ src/main/java/mod/gottsch/forge/claimmyland/core/item/CitizenDeed.java
@@ -104,6 +104,7 @@ public class CitizenDeed extends Deed {
         // if claiming an existing citizen parcel
         if (registryParcel.isPresent()) {
             if (registryParcel.get().getType() == ParcelType.CITIZEN) {
+                ClaimMyLand.LOGGER.debug("populating citizen foundation stone; placing player -> {}", String.valueOf(player));
                 // unwrap
                 CitizenParcel citizenParcel = (CitizenParcel) registryParcel.get();
                 // update block entity with properties of that of the existing citizen parcel
@@ -111,6 +112,7 @@ public class CitizenDeed extends Deed {
                 blockEntity.setNationEstateId(citizenParcel.getNationEstate().getId());
                 blockEntity.setRelativeBox(registryParcel.get().getSize());
                 blockEntity.setCoords(registryParcel.get().getCoords());
+                blockEntity.setPlacingPlayerId(player.getUUID());
             }
         }
     }